macosx playlist: fix issue with "expand triangle" alignment.
Only the 1st column in the NIB file has the expand triangle. If you want to move the triangle, you need to move ANOTHER column in front of the original "first column". The Playlistview was probably recreated at a time, but if you don't know this trick it's very difficult to restore that triangle to the position that we had been using.

This solves bug #2783