RTMP handshake compliance for clients
The spec says that the 2nd round of packets need to have the 1st 4 bytes be the
timestamp (potentially zero) and the next 4 bytes be zero. The bytes after that
should be random. (not necessarily cryptographically secure, just for
establishing a unique echo response in the 3rd round of packets).

Signed-off-by: Rémi Duraffort <ivoire@videolan.org>
(cherry picked from commit af0ba1445194d0251a6e8a0811cb15e83e6ef99e)

Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>